requirements.txt
langchain==1.2.6
langchain-openai==1.1.7
langgraph==1.0.6
langsmith==0.6.4
python-dotenv==1.2.1
pydantic==2.12.5
